requirements.txt
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
charset-normalizer==3.4.7
filelock==3.19.1
idna==3.11
lxml==6.0.4
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
snscrape==0.7.0.20230622
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
vk_api==11.10.0
